chore: trim unused components, hooks, deps
