--- components/mdx/Callout.tsx
@@ -1,3 +1,0 @@
-export default function Callout({ children }) {
-  return <div className="border-l-4 p-4">{children}</div>
-}
