--- components/mdx/DemoGrid.tsx
@@ -1,3 +1,0 @@
-export const DemoGrid = ({ children }) => (
-  <div className="demo-grid relative w-full grid p-0 gap-10 select-none">{children}</div>
-)
