+++ package.json
@@ -12,12 +12,7 @@
   "author": "Poimandres",
   "license": "MIT",
   "dependencies": {
-    "@codesandbox/sandpack-react": "^0.9.1",
     "@headlessui/react": "^1.2.0",
-    "@react-spring/web": "^9.3.1",
-    "@react-three/drei": "^7.0.9",
-    "@react-three/fiber": "^7.0.21",
-    "@react-three/postprocessing": "^2.0.5",
     "@sindresorhus/slugify": "^2.1.0",
     "@tailwindcss/aspect-ratio": "^0.3.0",
     "@tailwindcss/typography": "^0.4.1",
@@ -27,20 +22,13 @@
     "next": "^12.0.4",
     "next-mdx-remote": "^3.0.2",
     "node-fetch": "^3.1.0",
-    "postprocessing": "^6.22.0",
     "pretty-quick": "^3.1.1",
-    "prismjs": "^1.23.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-lazyload": "^3.1.0",
-    "react-prism": "^4.3.2",
     "react-spring": "9.3.2",
-    "react-twitter-widgets": "^1.10.0",
     "recursive-readdir": "^2.2.2",
     "remark-prism": "^1.3.4",
     "set-value": "^4.0.0",
-    "three": "^0.135.0",
-    "threejs-meshline": "^2.0.12",
     "zustand": "^3.5.4"
   },
   "devDependencies": {
